Preliminarily unify tileset processing
